refactor(pagination): refactor SbPagination components
