+++ src/components/Pagination/components/index.js
@@ -1,0 +1,3 @@
+export * from './shared'
+export * from './containers'
+export * from './modifiers'
